update list errors, models

+++ src/controllers/lists/delete-list.ts
@@ -1,9 +1,21 @@
 import { RequestHandler } from "express";
+import { validationResult } from "express-validator";
 
 import db from "../../db";
+import { RequestErrors } from "../../models/error";
 
 const deleteList: RequestHandler<{ listId: number }> = async (req, res, next) => {
+  const { userId } = req.user;
+  const { listId } = req.params;
+  const reqError = new RequestErrors();
   try {
+    // validation errors
+    const errors = validationResult(req);
+    if (!errors.isEmpty()) {
+      return res.status(401).json({ errors: errors.array() });
+    }
+
+    // db query
     const { rows }: { rows: { id: number }[] } = await db.query(
       `
     DELETE FROM lists
@@ -11,12 +23,14 @@ const deleteList: RequestHandler<{ listId: number }> = async (req, res, next) =>
     AND "userId" = $2
     RETURNING id;
     `,
-      [req.params.listId, req.user.userId]
+      [listId, userId]
     );
+
+    // null result error
     if (!rows.length) {
       res.status(401);
       return next({
-        message: `Unable to delete list (id ${req.params.listId}). The list may no longer exist, or you may not be authorized to delete it.`,
+        message: reqError.nullResult(),
       });
     }
     res.json({ message: "OK" });
@@ -24,7 +38,7 @@ const deleteList: RequestHandler<{ listId: number }> = async (req, res, next) =>
     console.log(error);
     res.status(500);
     next({
-      message: `An error occurred attempting to delete list id ${req.params.listId}.`,
+      message: reqError.internalServer(),
     });
   }
 };
